requirements.txt
streamlit
pandas
python-pptx
pywin32
